Fundamentals of Algorithmic Problem Solving
3.1 Aim of the program: Write a program to implement Binary Search to give the position of
leftmost appearance of the element in the array being searched. Display the number of
comparisons made while searching.

3.2 Aim of the program: Write a menu driven program to sort a list of elements in ascending
order using Insertion Sort technique. The nature of the input data is choice based and a distinct
file is considered for each choice. The sorted elements can be stored in a separate output file.
After sorting display the content of the output file along with number of comparisons. Based on
the number of comparisons, conclude the input scenario is either best or worst case.